+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # ChangeLog - PHP TableOfContents (TOC)
 All notable changes to this project are documented in this file.
 
+## 1.1 - 2015-05-06
+
+### Changed
+- Replaced `simple_dom_parser` with `masterminds/html5` parser library
+
+### Fixed
+- The `MarkupFixer` class now correctly preserves whitespace in HTML, correcting issues with PRE tags
+
 ## 1.0 - 2014-12-30
+
 ### Added
 - Initial Version